requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=9.0.0
opencv-python>=4.5.0
requests>=2.28.0
tqdm>=4.64.0
pyyaml>=6.0
wandb>=0.15.0
accelerate>=0.20.0
faiss-cpu>=1.7.0
huggingface-hub>=0.15.0
beir>=1.0.0